+++ src/main/java/com/restaurante/reservasapp/services/impl/ReservaServiceImpl.java
@@ -3,6 +3,7 @@ package com.restaurante.reservasapp.services.impl;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.lang.Nullable;
 import org.springframework.stereotype.Service;
 
 import com.restaurante.reservasapp.Entity.ReservaEntity;
@@ -16,18 +17,22 @@ public class ReservaServiceImpl implements ReservaService {
     private ReservaRepository repo;
 
     @Override
+    @SuppressWarnings("null")
     public ReservaEntity guardarReserva(ReservaEntity reserva) {
         return repo.save(reserva);
     }
 
     @Override
-    public ReservaEntity obtenerReserva(String id) {
-        return repo.findById(id).orElse(null);
+    @Nullable
+    public ReservaEntity obtenerReserva(@Nullable String id) {
+        return repo.findById(id != null ? id : "").orElse(null);
     }
 
     @Override
-    public void eliminarReserva(String id) {
-        repo.deleteById(id);
+    public void eliminarReserva(@Nullable String id) {
+        if (id != null) {
+            repo.deleteById(id);
+        }
     }
 
     @Override
