+++ src/main/java/com/restaurante/reservasapp/services/impl/UsuarioServiceImpl.java
@@ -3,6 +3,7 @@ package com.restaurante.reservasapp.services.impl;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.lang.Nullable;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
@@ -28,13 +29,16 @@ public class UsuarioServiceImpl implements UsuarioService {
     }
 
     @Override
-    public UsuarioEntity obtenerUsuario(String id) {
-        return repo.findById(id).orElse(null);
+    @Nullable
+    public UsuarioEntity obtenerUsuario(@Nullable String id) {
+        return repo.findById(id != null ? id : "").orElse(null);
     }
 
     @Override
-    public void eliminarUsuario(String id) {
-        repo.deleteById(id);
+    public void eliminarUsuario(@Nullable String id) {
+        if (id != null) {
+            repo.deleteById(id);
+        }
     }
 
     @Override
